# Do not use regex for host.Parse v2,v3 names