--- a/host/host_test.go
+++ b/host/host_test.go
@@ -419,3 +419,19 @@ func TestName_StringAll(t *testing.T) {
 		})
 	}
 }
+
+func BenchmarkParse(b *testing.B) {
+	// run the Fib function b.N times
+	names := []string{
+		"ndt-mlab1-foo01.mlab-sandbox.measurement-lab.org",
+		"ndt-lol12345-abcdef01.mlab.sandbox.measurement-lab.org",
+	}
+	for n := 0; n < b.N; n++ {
+		for i := range names {
+			_, err := Parse(names[i])
+			if err != nil {
+				panic(err)
+			}
+		}
+	}
+}