After reverting changes that incorporated the v3 hostname parsing into https://github.com/m-lab/locate/pull/179, we found that even the v2 name parsing appear to require significant CPU and memory resources owing to the use of regexp package.

This change eliminates the use of REGEX for v2 and v3 naming. This change preserves v1 support using regex since there are no known uses of v1 today. This logic may be safely removed in a future change, or preserved for parsing historical data.

<!-- Reviewable:start -->
- - -
This change is [<img src="https://reviewable.io/review_button.svg" height="34" align="absmiddle" alt="Reviewable"/>](https://reviewable.io/reviews/m-lab/go/180)
<!-- Reviewable:end -->