@stephen-soltesz 2024-02-09
Before:
```
goos: darwin
goarch: amd64
pkg: github.com/m-lab/go/host
cpu: Intel(R) Core(TM) i9-9900K CPU @ 3.60GHz
BenchmarkParse
BenchmarkParse-16    	   19474	     61246 ns/op	  101116 B/op	     586 allocs/op
PASS
ok  	github.com/m-lab/go/host	1.946s
```

After:
```
goos: darwin
goarch: amd64
pkg: github.com/m-lab/go/host
cpu: Intel(R) Core(TM) i9-9900K CPU @ 3.60GHz
BenchmarkParse
BenchmarkParse-16    	 2304636	       515.1 ns/op	     304 B/op	       7 allocs/op
PASS
ok  	github.com/m-lab/go/host	1.841s
```